feat: merge code 3.17.0

+++ meeting-qt/meeting/meeting-ui-sdk/components/macx_helpers.h
@@ -31,7 +31,7 @@ public slots:
     bool openFolder(const QString& folder);
     int getDisplayId(int screenIndex);
     QString getDeviceName();
-    int getWindowId(WId wid);
+    uintptr_t getWindowId(WId wid);
     void hideTitleBar(QQuickWindow* window);
     static void setAppPolicy();
     void activeWindow(QWindow* window) const;
